Refactored query builder to use Q objects and build query in the main function
